refactor(#15): 회원가입 로직 리팩토링
